# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AdminUser {
  id: string;
  email: string;
  name: string;
}

export interface LoginResponse {
  token: string;
  admin: AdminUser;
}

export type HubStatus = (typeof HubStatus)[keyof typeof HubStatus];

export const HubStatus = {
  Active: "Active",
  Inactive: "Inactive",
} as const;

export interface Hub {
  id: string;
  name: string;
  location: string;
  serviceAreas: string[];
  status: HubStatus;
  createdAt: string;
}

export interface HubResponse {
  hub: Hub;
}

export interface HubsListResponse {
  hubs: Hub[];
  total: number;
}

export type CreateHubRequestStatus =
  (typeof CreateHubRequestStatus)[keyof typeof CreateHubRequestStatus];

export const CreateHubRequestStatus = {
  Active: "Active",
  Inactive: "Inactive",
} as const;

export interface CreateHubRequest {
  name: string;
  location: string;
  serviceAreas: string[];
  status: CreateHubRequestStatus;
}

export type UpdateHubRequestStatus =
  (typeof UpdateHubRequestStatus)[keyof typeof UpdateHubRequestStatus];

export const UpdateHubRequestStatus = {
  Active: "Active",
  Inactive: "Inactive",
} as const;

export interface UpdateHubRequest {
  name?: string;
  location?: string;
  serviceAreas?: string[];
  status?: UpdateHubRequestStatus;
}

export interface DeleteResponse {
  message: string;
}

export interface HubStatsResponse {
  total: number;
  active: number;
  inactive: number;
  totalServiceAreas: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetHubsParams = {
  status?: GetHubsStatus;
};

export type GetHubsStatus = (typeof GetHubsStatus)[keyof typeof GetHubsStatus];

export const GetHubsStatus = {
  Active: "Active",
  Inactive: "Inactive",
} as const;