# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FishTokri Admin API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: hubs
    description: Hub management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Super Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /hubs:
    get:
      operationId: getHubs
      tags: [hubs]
      summary: Get all hubs
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Active, Inactive]
      responses:
        "200":
          description: List of hubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubsListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createHub
      tags: [hubs]
      summary: Create a new hub
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHubRequest"
      responses:
        "201":
          description: Hub created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /hubs/{id}:
    get:
      operationId: getHub
      tags: [hubs]
      summary: Get a hub by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubResponse"
        "404":
          description: Hub not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateHub
      tags: [hubs]
      summary: Update a hub
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHubRequest"
      responses:
        "200":
          description: Hub updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubResponse"
        "404":
          description: Hub not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteHub
      tags: [hubs]
      summary: Delete a hub
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Hub not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /hubs/{id}/toggle-status:
    patch:
      operationId: toggleHubStatus
      tags: [hubs]
      summary: Toggle hub active/inactive status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hub status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubResponse"
        "404":
          description: Hub not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /hubs/stats/summary:
    get:
      operationId: getHubStats
      tags: [hubs]
      summary: Get hub statistics summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Hub statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubStatsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"
      required:
        - token
        - admin

    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
      required:
        - id
        - email
        - name

    Hub:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [Active, Inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - serviceAreas
        - status
        - createdAt

    HubResponse:
      type: object
      properties:
        hub:
          $ref: "#/components/schemas/Hub"
      required:
        - hub

    HubsListResponse:
      type: object
      properties:
        hubs:
          type: array
          items:
            $ref: "#/components/schemas/Hub"
        total:
          type: integer
      required:
        - hubs
        - total

    CreateHubRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [Active, Inactive]
      required:
        - name
        - location
        - serviceAreas
        - status

    UpdateHubRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [Active, Inactive]

    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    HubStatsResponse:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        inactive:
          type: integer
        totalServiceAreas:
          type: integer
      required:
        - total
        - active
        - inactive
        - totalServiceAreas

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message