# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Super Admin login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  admin: zod.object({
    id: zod.string(),
    email: zod.string(),
    name: zod.string(),
  }),
});

/**
 * @summary Get all hubs
 */
export const GetHubsQueryParams = zod.object({
  status: zod.enum(["Active", "Inactive"]).optional(),
});

export const GetHubsResponse = zod.object({
  hubs: zod.array(
    zod.object({
      id: zod.string(),
      name: zod.string(),
      location: zod.string(),
      serviceAreas: zod.array(zod.string()),
      status: zod.enum(["Active", "Inactive"]),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Create a new hub
 */
export const CreateHubBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  serviceAreas: zod.array(zod.string()),
  status: zod.enum(["Active", "Inactive"]),
});

/**
 * @summary Get a hub by ID
 */
export const GetHubParams = zod.object({
  id: zod.coerce.string(),
});

export const GetHubResponse = zod.object({
  hub: zod.object({
    id: zod.string(),
    name: zod.string(),
    location: zod.string(),
    serviceAreas: zod.array(zod.string()),
    status: zod.enum(["Active", "Inactive"]),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Update a hub
 */
export const UpdateHubParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateHubBody = zod.object({
  name: zod.string().optional(),
  location: zod.string().optional(),
  serviceAreas: zod.array(zod.string()).optional(),
  status: zod.enum(["Active", "Inactive"]).optional(),
});

export const UpdateHubResponse = zod.object({
  hub: zod.object({
    id: zod.string(),
    name: zod.string(),
    location: zod.string(),
    serviceAreas: zod.array(zod.string()),
    status: zod.enum(["Active", "Inactive"]),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Delete a hub
 */
export const DeleteHubParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteHubResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Toggle hub active/inactive status
 */
export const ToggleHubStatusParams = zod.object({
  id: zod.coerce.string(),
});

export const ToggleHubStatusResponse = zod.object({
  hub: zod.object({
    id: zod.string(),
    name: zod.string(),
    location: zod.string(),
    serviceAreas: zod.array(zod.string()),
    status: zod.enum(["Active", "Inactive"]),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get hub statistics summary
 */
export const GetHubStatsResponse = zod.object({
  total: zod.number(),
  active: zod.number(),
  inactive: zod.number(),
  totalServiceAreas: zod.number(),
});