# lib/api-zod/src/generated/types/createHubRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHubRequestStatus } from "./createHubRequestStatus";

export interface CreateHubRequest {
  name: string;
  location: string;
  serviceAreas: string[];
  status: CreateHubRequestStatus;
}