# lib/api-zod/src/generated/types/hub.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HubStatus } from "./hubStatus";

export interface Hub {
  id: string;
  name: string;
  location: string;
  serviceAreas: string[];
  status: HubStatus;
  createdAt: Date;
}