# lib/api-zod/src/generated/types/hubResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Hub } from "./hub";

export interface HubResponse {
  hub: Hub;
}