# lib/api-zod/src/generated/types/hubStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HubStatsResponse {
  total: number;
  active: number;
  inactive: number;
  totalServiceAreas: number;
}