# lib/api-zod/src/generated/types/hubsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Hub } from "./hub";

export interface HubsListResponse {
  hubs: Hub[];
  total: number;
}