# lib/api-zod/src/generated/types/updateHubRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FishTokri Admin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateHubRequestStatus } from "./updateHubRequestStatus";

export interface UpdateHubRequest {
  name?: string;
  location?: string;
  serviceAreas?: string[];
  status?: UpdateHubRequestStatus;
}